backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
duckdb
pandas
cryptography
python-dotenv
langgraph
langchain-google-genai
PyJWT
pymysql
pg8000
